[WiP] Fix site scraping issues
- Fixed bug - vishvasnews loading more posts
- Added check to stop scraping site once encountered a url in db
- Refactored article parser
- Added check for invalid site to scraper
